Move data structures to separate file
