+++ internal/driver/common/structures_test.go
